+++ src/textual/widgets/selection_list.py
@@ -1,0 +1,8 @@
+from ._selection_list import (
+    MessageSelectionType,
+    Selection,
+    SelectionError,
+    SelectionType,
+)
+
+__all__ = ["MessageSelectionType", "Selection", "SelectionError", "SelectionType"]